daemon: use sigaction() to install child_handler()
In a future change, the flags used for processing SIGCHLD will need to
be updated, which is only possible by using sigaction().

Replace signal() with an equivalent invocation of sigaction(), which
has the added benefit of using BSD semantics reliably and therefore
not needing the rearming call in the signal handler.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>